feat: Add leader field to User model and Team model
This commit adds a new field, `leader`, to the User model and the Team model. The `leader` field in the User model establishes a one-to-one relationship with the Team model, indicating that a user can be the leader of a team. This change allows for better organization and management of teams within the application.